Simplify and cleanuo itkDataImage private types.
